Refactor worker packages for CGO isolation
