+++ internal/workers/embedworker/image_embed_worker.go
@@ -1,4 +1,4 @@
-package workers
+package embedworker
 
 import (
 	"context"
