+++ internal/workers/mediaworker/media_process_worker.go
@@ -1,4 +1,4 @@
-package workers
+package mediaworker
 
 import (
 	"context"
